# openapi.yaml
openapi: 3.0.3
info:
  title: RescueBite API
  description: |
    Food waste reduction platform. Restaurants list near-expiry food at a discount.
    
    **Key Features:**
    - Price decay state machine (Fresh → Discounted → Free → Compost)
    - Flash Auction — last 30 min before expiry, reverse auction (lowest bid wins)
    - Route Optimization — TSP-lite nearest neighbor algorithm for drivers
    - Geofencing — auto-complete delivery when driver enters 100m radius
    - Shelter role — charities receive free/compost food
  version: 2.0.0
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RouteOptimization:
      type: object
      properties:
        driverStart:
          type: object
          properties:
            latitude: { type: number }
            longitude: { type: number }
        totalStops: { type: integer }
        totalDistanceKm: { type: number }
        estimatedMinutes: { type: integer }
        algorithm: { type: string, example: "nearest-neighbor-tsp" }
        route:
          type: array
          items:
            type: object
            properties:
              type: { type: string, enum: [PICKUP, DELIVERY] }
              name: { type: string }
              address: { type: string }
              latitude: { type: number }
              longitude: { type: number }
              distanceFromPrev: { type: number, description: "km from previous stop" }
paths:
...
  /auctions:
    get:
      tags: [Auctions]
      summary: List active flash auctions (reverse auction — lowest bid wins)
      parameters:
        - { in: query, name: cursor, schema: { type: string } }
      responses:
        '200': { description: Active auctions sorted by expiry }
  /auctions/{id}:
    get:
      tags: [Auctions]
      summary: Get auction with all bids
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      responses:
        '200': { description: Auction detail with bids sorted lowest first }
  /auctions/{id}/bid:
    post:
      tags: [Auctions]
      summary: Place a bid — must be LOWER than current price (reverse auction)
      security: [{ BearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: { type: number, example: 500, description: "Must be lower than current listing price" }
      responses:
        '201': { description: Bid placed }
        '400': { description: Bid too high or auction ended }
        '403': { description: Customers only }
...
  /drivers:
    post:
      tags: [Drivers]
      summary: Create driver profile
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string, example: "+7 777 123 4567" }
      responses:
        '201': { description: Driver created (pending approval) }
  /drivers/me:
    get:
      tags: [Drivers]
      summary: Get driver profile
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Driver profile }
  /drivers/me/status:
    patch:
      tags: [Drivers]
      summary: Toggle online/offline + update coordinates
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isOnline: { type: boolean }
                latitude: { type: number }
                longitude: { type: number }
      responses:
        '200': { description: Status updated }
  /drivers/me/location:
    patch:
      tags: [Drivers]
      summary: "🌍 GEOFENCING — Update location. Auto-completes delivery if within 100m of destination"
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [latitude, longitude]
              properties:
                latitude: { type: number, example: 43.238 }
                longitude: { type: number, example: 76.945 }
      responses:
        '200':
          description: Location updated + geofence check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  location: { type: object }
                  geofenceChecked: { type: integer }
                  autoCompleted: { type: array, items: { type: object } }
                  message: { type: string }
  /drivers/me/route:
    get:
      tags: [Drivers]
      summary: "🗺️ ROUTE OPTIMIZATION — TSP-lite nearest neighbor algorithm"
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: orderIds
          required: true
          schema: { type: string }
          description: "Comma-separated order IDs, e.g. uuid1,uuid2,uuid3"
          example: "550e8400-e29b-41d4-a716-446655440000,550e8400-e29b-41d4-a716-446655440001"
      responses:
        '200':
          description: Optimized route with stops in order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteOptimization' }
        '400': { description: No orderIds provided or location not set }
...
  /shelters:
    get:
      tags: [Shelters]
      summary: List approved shelters (public)
      responses:
        '200': { description: Shelter list }
    post:
      tags: [Shelters]
      summary: Create shelter profile
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, address, latitude, longitude]
              properties:
                name: { type: string, example: "Almaty Food Bank" }
                address: { type: string }
                latitude: { type: number, example: 43.250 }
                longitude: { type: number, example: 76.920 }
                capacity: { type: integer, default: 50 }
      responses:
        '201': { description: Shelter created (pending approval) }
  /shelters/me:
    get:
      tags: [Shelters]
      summary: Get shelter profile
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Shelter profile }
    patch:
      tags: [Shelters]
      summary: Update shelter profile
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Updated }
  /ingredients:
    get:
      tags: [Ingredients]
      summary: List all ingredients
      parameters:
        - { in: query, name: search, schema: { type: string } }
      responses:
        '200': { description: Ingredients with allergen tags }
    post:
      tags: [Ingredients]
      summary: Create ingredient (Admin or Vendor)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, example: "Salmon" }
                allergens: { type: array, items: { type: string }, example: [FISH] }
      responses:
        '201': { description: Created }
  /admin/vendors/pending:
    get:
      tags: [Admin]
      summary: List pending vendors
      security: [{ BearerAuth: [] }]
      responses: { '200': { description: Pending vendors } }
  /admin/vendors/{id}/approve:
    patch:
      tags: [Admin]
      summary: Approve vendor
      security: [{ BearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      responses: { '200': { description: Approved } }
  /admin/drivers/pending:
    get:
      tags: [Admin]
      summary: List pending drivers
      security: [{ BearerAuth: [] }]
      responses: { '200': { description: Pending drivers } }
  /admin/drivers/{id}/approve:
    patch:
      tags: [Admin]
      summary: Approve driver
      security: [{ BearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      responses: { '200': { description: Approved } }
  /admin/shelters/pending:
    get:
      tags: [Admin]
      summary: List pending shelters
      security: [{ BearerAuth: [] }]
      responses: { '200': { description: Pending shelters } }
  /admin/shelters/{id}/approve:
    patch:
      tags: [Admin]
      summary: Approve shelter
      security: [{ BearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      responses: { '200': { description: Approved } }
  /admin/users/{id}/suspend:
    patch:
      tags: [Admin]
      summary: Suspend user
      security: [{ BearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses: { '200': { description: Suspended } }
  /admin/analytics:
    get:
      tags: [Admin]
      summary: Platform analytics
      security: [{ BearerAuth: [] }]
      responses: { '200': { description: Analytics data } }
  /admin/refunds/pending:
    get:
      tags: [Admin]
      summary: List pending refunds
      security: [{ BearerAuth: [] }]
      responses: { '200': { description: Refund requests } }
  /admin/refunds/{id}/resolve:
    patch:
      tags: [Admin]
      summary: Resolve refund
      security: [{ BearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                resolution: { type: string, enum: [REFUNDED, REJECTED] }
      responses: { '200': { description: Resolved } }